notebooks/requirements.txt
albumentations
jupyter
numpy
pandas
Pillow
scipy
tqdm
torchviz
torch-summary
pytorch-lightning
scikit-learn
matplotlib
seaborn
pydotplus